remove warning

+++ srtparser.h
@@ -259,7 +259,7 @@ inline SubRipParser::SubRipParser(std::string fileName, bool keepHTML, bool doNo
 
 inline SubRipParser::~SubRipParser(void)
 {
-    for(int i=0;i != _subtitles.size();++i)
+    for(size_t i=0;i != _subtitles.size();++i)
     {
         if(_subtitles[i])
             delete _subtitles[i];
